move constants into constants file to keep useTextEditor file clean
